+++ src/test.rs
@@ -23,7 +23,7 @@ fn test_struct()
     assert_eq!
     (
         const_definition!(#[derive(Debug)] TestStruct), 
-        format!("#[derive(Debug)]struct TestStruct{{ test_u8: u8, test_vec: &\'static [&\'static str], }}")
+        format!("#[derive(Debug)] struct TestStruct{{ test_u8: u8, test_vec: &\'static [&\'static str], }}")
     );
     assert_eq!
     (
@@ -38,7 +38,7 @@ fn test_struct_definition()
     assert_eq!
     (
         const_definition!(#[derive(Debug)] TestStruct), 
-        format!("#[derive(Debug)]struct TestStruct{{ test_u8: u8, test_vec: &\'static [&\'static str], }}")
+        format!("#[derive(Debug)] struct TestStruct{{ test_u8: u8, test_vec: &\'static [&\'static str], }}")
     );
 }
 
@@ -107,8 +107,8 @@ fn test_strings()
 {
     assert_eq!
     (
-        const_declaration!(TEST_STR = "I'm a string!"),
-        format!("const TEST_STR: &'static str = \"I'm a string!\";")
+        const_declaration!(pub(in crate) TEST_STR = "I'm a string!"),
+        format!("pub(in crate)  const TEST_STR: &'static str = \"I'm a string!\";")
     );
     assert_eq!
     (
@@ -129,8 +129,8 @@ fn test_nums()
     {
         assert_eq!
         (
-            val.const_declaration(var_name), 
-            format!("const {0}: {1} = {2}{1};", var_name, type_name, val)
+            val.const_declaration("pub", var_name), 
+            format!("pub const {0}: {1} = {2}{1};", var_name, type_name, val)
         );
     }
     test("TEST_U8", "u8", u8::MAX);